React story and compound item
