+++ packages/react/src/UnorderedList/index.ts
@@ -1,2 +1,1 @@
 export { UnorderedList } from './UnorderedList'
-export { UnorderedListItem } from './UnorderedListItem'
